% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{centralTend}
\alias{centralTend}
\title{Central Tendency}
\usage{
centralTend(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Calculates the central tendency of a given variable i.e., Mean and median
}
\examples{
centralTend(x)
}
