% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{decisionTree}
\alias{decisionTree}
\title{Decison Tree}
\usage{
decisionTree(t, d)
}
\arguments{
\item{t}{Target Variable}

\item{d}{Dataset}
}
\value{

}
\description{
Generates a Decision tree plot with two branches at each value to predict the output on the basis of given dataset
}
\note{
Pass the target value with '$' .
Ex- dataset$target(d$t)
}
\examples{
decisionTree(t,d)
}
