% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{My hello World Function}
\usage{
hello(x)
}
\arguments{
\item{x}{The name of person to say hi to}
}
\value{
The output from \code{\link{print}}
}
\description{
My hello World Function
}
\examples{
hello("Nishant")
\dontrun{
hello("Steve")
}
}
