% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hierarchicalClust}
\alias{hierarchicalClust}
\title{Hierarchical Clustering}
\usage{
hierarchicalClust(t, d)
}
\arguments{
\item{t}{The Target Variable}

\item{d}{The Dataset}
}
\value{

}
\description{
Performs Hierarchical Clustering algorithm on target variable and given dataset
Also constructs a dendogram plot depicting the merging of clusters
}
\note{
Pass the target value with '$' .
Ex- dataset$target(d$t)
}
\examples{
hierarchicalClust(t,d)
}
