% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{kmeansClust}
\alias{kmeansClust}
\title{K means Clustering}
\usage{
kmeansClust(x, k)
}
\arguments{
\item{x}{The Variable}

\item{k}{Number of Clusters}
}
\value{

}
\description{
Performs K means clustering algorithm and also develops stepwise plots of how clusters are being formed
}
\examples{
kmeansClust(x,k)
}
