% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{knearestNeighbour}
\alias{knearestNeighbour}
\title{K nearest Neighbours}
\usage{
knearestNeighbour(t, d)
}
\arguments{
\item{t}{Target Variable}

\item{d}{Dataset}
}
\value{

}
\description{
Performs knn classification on a target variable and a given dataset and predicts the accuracy by building a confusion matrix
The training and testing data is splitted in the ration of 80:20 respectively
}
\note{
Pass the target value with '$' .
Ex- dataset$target(d$t)
}
\examples{
knearestNeighbour(t,d)
}
