% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{linearRegression}
\alias{linearRegression}
\title{Linear Regression Algorithm}
\usage{
linearRegression(y, x)
}
\arguments{
\item{y}{}

\item{x}{}
}
\value{

}
\description{
Performs Linear on the two input variables and
plots the line of best fit between them
}
\examples{
linearRegression(y,x)
}
