% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{logisticRegression}
\alias{logisticRegression}
\title{Logistic Regression algorithm for classification}
\usage{
logisticRegression(t, d)
}
\arguments{
\item{t}{Target Variable}

\item{d}{Dataset

 @note Pass the target value with '$' .
Ex- dataset$target(d$t)}
}
\value{

}
\description{
Predicts the probability of the target variable on the basis of input dataset by performing Logistic Regression Algorithm and
also Plots a curve between 0 and 1 in the y axis
}
\examples{
logisticRegression(t,d)
}
