\name{mlearning-package}
\alias{mlearning-package}
\docType{package}

\title{mlearning: machine learning algorithms with a common UI and confusion matrices}
\description{
  This package provides wrappers around several existing machine learning
  algorithms in R, under a unified user interface. Confusion matrices can also
  be calculated and viewed as tables or plots.
}

\details{
  \tabular{ll}{
    Package: \tab mlearning\cr
    Type: \tab Package\cr
    Version: \tab 1.0-0\cr
    Date: \tab 2012-08-04\cr
    License: \tab GPL 2 or above at your convenience.\cr
  }
}

\author{
  Philippe Grosjean & Kevin Denis, Numerical Ecology of Aquatic Systems,
  Mons University, Belgium.

  Maintainer: Philippe Grosjean <Philippe.Grosjean@umons.ac.be>
}

%\references{
%  Not yet!
%}

\seealso{ \code{\link{mlearning}}, \code{\link{confusion}} }

%\examples{
%# TODO...
%}

\keyword{package}
