% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{multipleRegression}
\alias{multipleRegression}
\title{Multiple Regression Algortihm}
\usage{
multipleRegression(t, d)
}
\arguments{
\item{t}{Target Variable}

\item{d}{Dataset}
}
\value{

}
\description{
Performs the Multiple linear Regression between the input y value and all other columns variables in the given dataset
}
\note{
Pass the target value with '$' .
Ex- dataset$target(d$t)
}
\examples{
multipleRegression(y,d)
}
