% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{quartiles}
\alias{quartiles}
\title{Quartiles}
\usage{
quartiles(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Calculates the percentage quartile density of the variable and
Interquartile range of the given Variable
}
\examples{
quartiles(x)
}
