% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train}
\alias{train}
\alias{train.default}
\title{Get the training variable for a mlearning object}
\usage{
train(object, ...)

\method{train}{default}(object, ...)
}
\arguments{
\item{object}{an object having a train attribute.}

\item{...}{further parameter (depends on the method).}
}
\value{
A data frame containing the training variables of the model.
}
\description{
The training variables (train) are the variables used to train a classifier,
excepted the prediction (class or dependent variable).
}
\examples{
data("HouseVotes84", package = "mlbench")
house_rf <- ml_rforest(data = HouseVotes84, Class ~ .)
house_rf
train(house_rf)
}
\seealso{
\code{\link[=mlearning]{mlearning()}}, \code{\link[=response]{response()}}, \code{\link[=confusion]{confusion()}}
}
