% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{variability}
\alias{variability}
\title{Variability}
\usage{
variability(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Calculates the variability for a given
variable i.e., Minimum , Maximum , Range , Standard Deviation
Also develops a Barplot to visualize the variability of the variable
}
\examples{
variability(x)
}
